requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
mysql-connector-python>=9.2.0
openpyxl>=3.1.0
